- separate onBoarding item and pageViewWidget to clean onBoarding Screen
